# cmd/laas/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Audit:
    properties:
      id:
        example: 456
        type: integer
      timestamp:
        example: "2023-12-01T18:10:25.00+05:30"
        type: string
      type:
        example: license
        type: string
      type_id:
        example: 34
        type: integer
      user_id:
        example: 123
        type: integer
    type: object
  models.AuditResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Audit'
        type: array
      paginationmeta:
        $ref: '#/definitions/models.PaginationMeta'
      status:
        example: 200
        type: integer
    type: object
...
  models.LicenseError:
    properties:
      error:
        example: invalid request body
        type: string
      message:
        example: invalid request body
        type: string
      path:
        example: /api/v1/licenses
        type: string
      status:
        example: 400
        type: integer
      timestamp:
        example: "2023-12-01T10:00:51+05:30"
        type: string
    type: object
...
  models.NullString:
    properties:
      string:
        type: string
      valid:
        description: Valid is true if String is not NULL
        type: boolean
    type: object
  models.Obligation:
    properties:
      active:
        type: boolean
      classification:
        enum:
        - green
        - white
        - yellow
        - red
        example: green
        type: string
      comment:
        $ref: '#/definitions/models.NullString'
      id:
        example: 147
        type: integer
      modifications:
        example: true
        type: boolean
      text:
        example: Source code be made available when distributing the software.
        type: string
      text_updatable:
        example: true
        type: boolean
      topic:
        example: copyleft
        type: string
      type:
        enum:
        - obligation
        - restriction
        - risk
        - right
        example: risk
        type: string
    type: object
...
  models.ObligationPATCHRequestJSONSchema:
    type: object
  models.ObligationPOSTRequestJSONSchema:
    properties:
      active:
        example: true
        type: boolean
      classification:
        enum:
        - green
        - white
        - yellow
        - red
        type: string
      comment:
        $ref: '#/definitions/models.NullString'
      modifications:
        type: boolean
      shortnames:
        example:
        - GPL-2.0-only
        - GPL-2.0-or-later
        items:
          type: string
        type: array
      text:
        example: Source code be made available when distributing the software.
        type: string
      topic:
        example: copyleft
        type: string
      type:
        enum:
        - obligation
        - restriction
        - risk
        - right
        type: string
    required:
    - active
    - classification
    - comment
    - modifications
    - shortnames
    - text
    - topic
    - type
    type: object
  models.ObligationResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Obligation'
        type: array
      paginationmeta:
        $ref: '#/definitions/models.PaginationMeta'
      status:
        example: 200
        type: integer
    type: object
...
  models.PaginationMeta:
    properties:
      limit:
        example: 10
        type: integer
      next:
        example: /api/v1/licenses?limit=10&page=11
        type: string
      page:
        example: 10
        type: integer
      previous:
        example: /api/v1/licenses?limit=10&page=9
        type: string
      resource_count:
        example: 200
        type: integer
      total_pages:
        example: 20
        type: integer
    type: object
...
info:
  contact:
    email: fossology@fossology.org
    name: FOSSology
    url: https://fossology.org
  description: Service to host license information for other services to query over
    REST API.
  license:
    name: GPL-2.0-only
    url: https://github.com/fossology/LicenseDb/blob/main/LICENSE
  title: laas (License as a Service) API
  version: 0.0.9
paths:
...
  /obligations:
    get:
      consumes:
      - application/json
      description: Get all active obligations from the service
      operationId: GetAllObligation
      parameters:
      - description: Active obligation only
        in: query
        name: active
        required: true
        type: boolean
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of records per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ObligationResponse'
        "404":
          description: No obligations in DB
          schema:
            $ref: '#/definitions/models.LicenseError'
      summary: Get all active obligations
      tags:
      - Obligations
    post:
      consumes:
      - application/json
      description: Create an obligation and associate it with licenses
      operationId: CreateObligation
      parameters:
      - description: Obligation to create
        in: body
        name: obligation
        required: true
        schema:
          $ref: '#/definitions/models.ObligationPOSTRequestJSONSchema'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ObligationResponse'
        "400":
          description: Bad request body
          schema:
            $ref: '#/definitions/models.LicenseError'
        "409":
          description: Obligation with same body exists
          schema:
            $ref: '#/definitions/models.LicenseError'
        "500":
          description: Unable to create obligation
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - ApiKeyAuth: []
      summary: Create an obligation
      tags:
      - Obligations
  /obligations/{topic}:
    delete:
      consumes:
      - application/json
      description: Deactivate an obligation
      operationId: DeleteObligation
      parameters:
      - description: Topic of the obligation to be updated
        in: path
        name: topic
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: No obligation with given topic found
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - ApiKeyAuth: []
      summary: Deactivate obligation
      tags:
      - Obligations
    get:
      consumes:
      - application/json
      description: Get an active based on given topic
      operationId: GetObligation
      parameters:
      - description: Topic of the obligation
        in: path
        name: topic
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ObligationResponse'
        "404":
          description: No obligation with given topic found
          schema:
            $ref: '#/definitions/models.LicenseError'
      summary: Get an obligation
      tags:
      - Obligations
    patch:
      consumes:
      - application/json
      description: Update an existing obligation record
      operationId: UpdateObligation
      parameters:
      - description: Topic of the obligation to be updated
        in: path
        name: topic
        required: true
        type: string
      - description: Obligation to be updated
        in: body
        name: obligation
        required: true
        schema:
          $ref: '#/definitions/models.ObligationPATCHRequestJSONSchema'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ObligationResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.LicenseError'
        "404":
          description: No obligation with given topic found
          schema:
            $ref: '#/definitions/models.LicenseError'
        "500":
          description: Unable to update obligation
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - ApiKeyAuth: []
      summary: Update obligation
      tags:
      - Obligations
  /obligations/{topic}/audits:
    get:
      consumes:
      - application/json
      description: Fetches audits corresponding to an obligation
      operationId: GetObligationAudits
      parameters:
      - description: Topic of the obligation for which audits need to be fetched
        in: path
        name: topic
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of records per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuditResponse'
        "404":
          description: No obligation with given topic found
          schema:
            $ref: '#/definitions/models.LicenseError'
        "500":
          description: unable to find audits with such obligation topic
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - ApiKeyAuth: []
      summary: Fetches audits corresponding to an obligation
      tags:
      - Obligations
...
securityDefinitions:
  ApiKeyAuth:
    description: Token from /login endpoint
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"